Project Complete; Ready For Review
I have now completed the project in its entirety with documentation to boot. Code doesn't seem to be functioning that well, but it's much more productive off my time to just wrap this one up and move on then to stay stuck like this (which I've kind of already choose to before letting go on this).